requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
python-docx==0.8.11
cohere==4.32
Werkzeug==2.3.7
